+++ effectful-core/CHANGELOG.md
@@ -1,10 +1,10 @@
-# effectful-core-2.5.1.0 (????-??-??)
+# effectful-core-2.5.1.0 (2024-11-27)
 * Add `passthrough` to `Effectful.Dispatch.Dynamic` for passing operations to
   the upstream handler within `interpose` and `impose` without having to fully
   pattern match on them.
 * **Bugfixes**:
-  - Fix a potential space leak related to `HasCallStack` quirks
-    (https://gitlab.haskell.org/ghc/ghc/-/issues/25520).
+  - Fix a potential space leak related to `HasCallStack` quirks (see
+    https://gitlab.haskell.org/ghc/ghc/-/issues/25520 for more information).
 
 # effectful-core-2.5.0.0 (2024-10-23)
 * Add `plusEff` (specialized version of `<|>`) to `Effectful.NonDet` and make