+++ effectful-core/CHANGELOG.md
@@ -1,4 +1,4 @@
-# effectful-core-2.6.0.0 (????-??-??)
+# effectful-core-2.6.0.0 (2025-06-13)
 * Adjust `generalBracket` with `base >= 4.21` to make use of the new exception
   annotation mechanism.
 * Add `withException` to `Effectful.Exception`.